requirements.txt
Django==5.2.9
dj-database-url==3.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
Pillow==12.0.0
python-dotenv==1.2.1
gunicorn==23.0.0
whitenoise==6.11.0
